feat: lock projects from deletion

+++ src/dpx/cli/delete.py
@@ -6,6 +6,7 @@ rm
 """
 
 import shutil
+import threading
 from pathlib import Path
 from typing_extensions import Annotated
 
@@ -13,11 +14,12 @@ import typer
 from icecream import ic
 from rich import print
 
-from src.dpx.cli.utils.util import ProjectManager
+from src.dpx.cli.utils.util import ProjectManager, Project
 from src.dpx.utils.paths import PROJECTS_DIR
 
 app = typer.Typer()
 current_main = "main"
+wait_to_unlock: int = 10
 
 
 @app.command(help="Delete project(s).")
@@ -94,14 +96,29 @@ def rm(
         """Want to change to move to .trash/"""
         if len(filepaths) == 0:
             print("No files deleted.")
+            return
 
+        unlocked_projects: list[str] = []
         for f in filepaths:
-            try:
+            if not f.exists():
+                # file not found error
+                print(f"'{f.name}' not a project in '{f.parent.name}'")
+                continue
+
+            project = Project(f)
+            if project.is_locked():
+                # print(f"'{project.name}' was not deleted. Must be unlocked first.")
+                unlocked_projects.append(project.name)
+            else:
                 shutil.rmtree(f)
                 print(f"'{f.name}' deleted from '{f.parent.name}'")
-            except FileNotFoundError:
-                print(f"'{f.name}' not a project in '{f.parent.name}'")
-        pass
+                continue
+
+        if unlocked_projects:
+            plural = len(unlocked_projects) > 1
+            for p in unlocked_projects:
+                print(f"'{p}'", end=", " if plural else " ")
+            print(f"must be unlocked before {'they are' if plural else 'it is'} deleted.")
 
     if rm_all_temps:
         search_all = True
